RAG1_DEV_HANDOFF/requirements.txt
openai>=1.30
chromadb>=0.5
pymupdf>=1.24
pydantic>=2.0
python-dotenv>=1.0
numpy>=1.23
pydicom>=3.0
Pillow>=10.0
torch>=2.0
torchvision>=0.15
ultralytics>=8.4
fastapi>=0.115
uvicorn>=0.30
python-multipart>=0.0.9
